Moved unsupported test
