+++ drools-test-coverage/test-compiler-integration/src/test/java/org/drools/compiler/integrationtests/AccumulateTest.java
@@ -2935,111 +2935,6 @@ public class AccumulateTest {
         Assertions.assertThat(kieBuilder.getResults().getMessages()).isNotEmpty();
     }
 
-    @Test
-    public void testAccFunctionOpaqueJoins() {
-        // DROOLS-661
-        testAccFunctionOpaqueJoins(PropertySpecificOption.ALLOWED);
-    }
-
-    @Test
-    public void testAccFunctionOpaqueJoinsWithPropertyReactivity() {
-        // DROOLS-1445
-        testAccFunctionOpaqueJoins(PropertySpecificOption.ALWAYS);
-    }
-
-    private void testAccFunctionOpaqueJoins(final PropertySpecificOption propertySpecificOption) {
-        final String drl = "package org.test; " +
-                "import java.util.*; " +
-                "global List list; " +
-                "global List list2; " +
-
-                "declare Tick " +
-                "  tick : int " +
-                "end " +
-
-                "declare Data " +
-                "  values : List " +
-                "  bias : int = 0 " +
-                "end " +
-
-                "rule Init " +
-                "when " +
-                "then " +
-                "  insert( new Data( Arrays.asList( 1, 2, 3 ), 1 ) ); " +
-                "  insert( new Data( Arrays.asList( 4, 5, 6 ), 2 ) ); " +
-                "  insert( new Tick( 0 ) );" +
-                "end " +
-
-                "rule Update " +
-                "  no-loop " +
-                "when " +
-                "  $i : Integer() " +
-                "  $t : Tick() " +
-                "then " +
-                "  System.out.println( 'Set tick to ' + $i ); " +
-                "  modify( $t ) { " +
-                "      setTick( $i ); " +
-                "  } " +
-                "end " +
-
-                "rule M " +
-                "  dialect 'mvel' " +
-                "when " +
-                "    Tick( $index : tick ) " +
-                "    accumulate ( $data : Data( $bias : bias )," +
-                "                 $tot : sum( $data.values[ $index ] + $bias ) ) " +
-                "then " +
-                "    System.out.println( $tot + ' for J ' + $index ); " +
-                "    list.add( $tot ); " +
-                "end " +
-
-                "rule J " +
-                "when " +
-                "    Tick( $index : tick ) " +
-                "    accumulate ( $data : Data( $bias : bias )," +
-                "                 $tot : sum( ((Integer)$data.getValues().get( $index )) + $bias ) ) " +
-                "then " +
-                "    System.out.println( $tot + ' for M ' + $index ); " +
-                "    list2.add( $tot ); " +
-                "end ";
-
-        final ReleaseId releaseId1 = KieServices.get().newReleaseId("org.kie", "accumulate-test", "1");
-        final Map<String, String> kieModuleConfigurationProperties = new HashMap<>();
-        kieModuleConfigurationProperties.put(PropertySpecificOption.PROPERTY_NAME, propertySpecificOption.toString());
-
-        final KieModule kieModule = KieUtil.getKieModuleFromDrls(releaseId1,
-                                                                 kieBaseTestConfiguration,
-                                                                 KieSessionTestConfiguration.STATEFUL_REALTIME,
-                                                                 kieModuleConfigurationProperties,
-                                                                 drl);
-        final KieContainer kieContainer = KieServices.get().newKieContainer(kieModule.getReleaseId());
-        final KieBase kbase = kieContainer.getKieBase();
-        final KieSession ks = kbase.newKieSession();
-        try {
-            final List list = new ArrayList();
-            ks.setGlobal("list", list);
-            final List list2 = new ArrayList();
-            ks.setGlobal("list2", list2);
-
-            // init data
-            ks.fireAllRules();
-            assertEquals(Collections.singletonList(8.0), list);
-            assertEquals(Collections.singletonList(8.0), list2);
-
-            ks.insert(1);
-            ks.fireAllRules();
-            assertEquals(asList(8.0, 10.0), list);
-            assertEquals(asList(8.0, 10.0), list2);
-
-            ks.insert(2);
-            ks.fireAllRules();
-            assertEquals(asList(8.0, 10.0, 12.0), list);
-            assertEquals(asList(8.0, 10.0, 12.0), list2);
-        } finally {
-            ks.dispose();
-        }
-    }
-
     public static class ExpectedMessage {
 
         String type;
